GoogleKeywordIdeas/requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-ads==24.1.0
google-api-core==2.19.1
googleapis-common-protos==1.63.2
google-auth==2.31.0
google-auth-oauthlib==1.2.0
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
numpy==2.0.0
oauthlib==3.2.2
pandas==2.2.2
pip==22.3.1
protobuf==4.25.3
proto-plus==1.24.0
pyasn1_modules==0.4.0
pyasn1==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
